Add msbuild-based stage0 where possible
