+++ build_projects/dotnet-cli-build/DotNetPackPJ.cs
@@ -3,22 +3,24 @@
 
 namespace Microsoft.DotNet.Cli.Build
 {
-    public class DotNetPackMsbuildLegacy : DotNetTool
+    public class DotNetPackPJ : DotNetTool
     {
         protected override string Command
         {
-            get { return "pack3"; }
+            get { return "pack"; }
         }
 
         protected override string Args
         {
-            get { return $"{GetProjectPath()} {GetConfiguration()} {GetNoBuild()} {GetOutput()} {GetVersionSuffix()}"; }
+            get { return $"{GetProjectPath()} {GetConfiguration()} {GetNoBuild()} {GetBuildBasePath()} {GetOutput()} {GetVersionSuffix()}"; }
         }
 
         public string Configuration { get; set; }
 
         public bool NoBuild { get; set; }
-        
+
+        public string BuildBasePath { get; set; }
+
         public string Output { get; set; }
 
         public string ProjectPath { get; set; }
@@ -44,7 +46,17 @@ namespace Microsoft.DotNet.Cli.Build
 
             return null;
         }
-        
+
+        private string GetBuildBasePath()
+        {
+            if (!string.IsNullOrEmpty(BuildBasePath))
+            {
+                return $"--build-base-path {BuildBasePath}";
+            }
+
+            return null;
+        }
+
         private string GetOutput()
         {
             if (!string.IsNullOrEmpty(Output))
