+++ test/Microsoft.DotNet.Tools.Tests.Utilities/RepoDirectoriesProvider.cs
@@ -73,7 +73,7 @@ namespace Microsoft.DotNet.Tools.Test.Utilities
 
         private string GetPjDotnetPath()
         {
-            return new DirectoryInfo(Path.Combine(RepoRoot, ".dotnet_stage0"))
+            return new DirectoryInfo(Path.Combine(RepoRoot, ".dotnet_stage0PJ"))
                 .GetDirectories().First()
                 .GetFiles("dotnet*").First()
                 .FullName;
